requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=0.6,!=0.7,<1.0

Paste
PasteDeploy>=1.5.0
Routes>=1.12.3,!=2.0
eventlet>=0.15.2
greenlet>=0.3.2
httplib2>=0.7.5
requests>=2.2.0,!=2.4.0
jsonrpclib
Jinja2>=2.6  # BSD License (3 clause)
keystonemiddleware>=1.0.0
netaddr>=0.7.12
python-neutronclient>=2.3.6,<3
retrying>=1.2.3,!=1.3.0 # Apache-2.0
SQLAlchemy>=0.9.7,<=0.9.99
WebOb>=1.2.3
python-keystoneclient>=0.11.1
alembic>=0.7.1
six>=1.7.0
stevedore>=1.1.0  # Apache-2.0
oslo.config>=1.6.0  # Apache-2.0
oslo.context>=0.1.0                     # Apache-2.0
oslo.db>=1.3.0  # Apache-2.0
oslo.i18n>=1.0.0  # Apache-2.0
oslo.messaging>=1.4.0,!=1.5.0
oslo.middleware>=0.3.0                  # Apache-2.0
oslo.rootwrap>=1.3.0
oslo.serialization>=1.0.0               # Apache-2.0
oslo.utils>=1.2.0                       # Apache-2.0

python-novaclient>=2.18.0
pecan>=0.8.0
